+++ src/Resolution.test.ts
@@ -54,6 +54,32 @@ describe('Resolution', () => {
     expect(ethAddress).toBe('0x8aaD44321A86b170879d7A244c1e8d360c99DdA8');
   });
 
+  it('should resolve a custom record', async () => {
+    const resolution = new Resolution({blockchain: {cns: {url: secretInfuraLink()}}});
+    const customRecord = 'gundb.username.value';
+    const eyes = mockAsyncMethods(resolution.cns, {
+      getResolver: '0x878bC2f3f717766ab69C0A5f9A6144931E61AEd3',
+      getRecord: '0x47992daf742acc24082842752fdc9c875c87c56864fee59d8b779a91933b159e48961566eec6bd6ce3ea2441c6cb4f112d0eb8e8855cc9cf7647f0d9c82f00831c'
+    });
+    const value = await resolution.record("homecakes.crypto", customRecord);
+    expectSpyToBeCalled(eyes);
+    expect(value).toBe('0x47992daf742acc24082842752fdc9c875c87c56864fee59d8b779a91933b159e48961566eec6bd6ce3ea2441c6cb4f112d0eb8e8855cc9cf7647f0d9c82f00831c');
+  });
+
+  it('should error with recordNoFound for custom record', async () => {
+    const resolution = new Resolution({blockchain: {cns: {url: secretInfuraLink()}}});
+    const customWrongRecord = 'noSuchRecordEver';
+    const eyes = mockAsyncMethods(resolution.cns, {
+      getResolver: '0x878bC2f3f717766ab69C0A5f9A6144931E61AEd3',
+      getRecord: ''
+    });
+    await expectResolutionErrorCode(
+      resolution.record("homecakes.crypto", customWrongRecord), 
+      ResolutionErrorCode.RecordNotFound
+    );
+    expectSpyToBeCalled(eyes);
+  })
+
   it('checks Resolution#addressOrThrow error #1', async () => {
     const resolution = new Resolution();
     await expectResolutionErrorCode(